refactor: move login and register test classes to 'Public' folder
